Финальный проект №11 - рефакторинг - #1
