+++ src/main/resources/application.properties
@@ -3,3 +3,4 @@ server.port=8080
 #logging.level.org.springframework.web=TRACE
 logging.level.ru.yandex.practicum.filmorate.controller.FilmController = TRACE
 logging.level.ru.yandex.practicum.filmorate.controller.UserController = TRACE
+server.error.include-message=always
